refactor: organize components
