+++ src/lib/components/features/Stats.svelte
