+++ src/lib/components/features/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Feature Components index
+ * @module components/features
+ */
+
+export { default as Feed } from './Feed.svelte'
+export { default as Stats } from './Stats.svelte'
+export { default as ActionButton } from './ActionButton.svelte'
