+++ src/lib/components/index.ts
@@ -3,18 +3,25 @@
  * @module components
  */
 
-export { default as Button } from './Button.svelte'
-export { default as ErrorAlert } from './ErrorAlert.svelte'
-export { default as Feed } from './Feed.svelte'
-export { default as Input } from './Input.svelte'
-export { default as LoadingSpinner } from './LoadingSpinner.svelte'
+// UI Components
+export { default as Button } from './ui/Button.svelte'
+export { default as Input } from './ui/Input.svelte'
+export { default as Textarea } from './ui/Textarea.svelte'
+export { default as Select } from './ui/Select.svelte'
+export { default as Toggle } from './ui/Toggle.svelte'
+export { default as LoadingSpinner } from './ui/LoadingSpinner.svelte'
+export { default as ErrorAlert } from './ui/ErrorAlert.svelte'
+export { default as ThemeToggle } from './ui/ThemeToggle.svelte'
+export { default as ScrollArea } from './ui/ScrollArea.svelte'
+
+// Layout Components
+export { default as Modal } from './layout/Modal.svelte'
+export { default as Table } from './layout/Table.svelte'
+
+// Feature Components
+export { default as Feed } from './features/Feed.svelte'
+export { default as Stats } from './features/Stats.svelte'
+export { default as ActionButton } from './features/ActionButton.svelte'
+
+// Brand Components
 export { default as Logo } from './Logo.svelte'
-export { default as Modal } from './Modal.svelte'
-export { default as ScrollArea } from './ScrollArea.svelte'
-export { default as Select } from './Select.svelte'
-export { default as Stats } from './Stats.svelte'
-export { default as Table } from './Table.svelte'
-export { default as Textarea } from './Textarea.svelte'
-export { default as ThemeToggle } from './ThemeToggle.svelte'
-export { default as Toggle } from './Toggle.svelte'
-export { default as ActionButton } from './ActionButton.svelte'
