+++ src/lib/components/layout/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Layout Components index
+ * @module components/layout
+ */
+
+export { default as Modal } from './Modal.svelte'
+export { default as Table } from './Table.svelte'
