+++ src/lib/components/ui/ErrorAlert.svelte
