+++ src/lib/components/ui/index.ts
@@ -1,0 +1,14 @@
+/**
+ * UI Components index
+ * @module components/ui
+ */
+
+export { default as Button } from './Button.svelte'
+export { default as Input } from './Input.svelte'
+export { default as Textarea } from './Textarea.svelte'
+export { default as Select } from './Select.svelte'
+export { default as Toggle } from './Toggle.svelte'
+export { default as LoadingSpinner } from './LoadingSpinner.svelte'
+export { default as ErrorAlert } from './ErrorAlert.svelte'
+export { default as ThemeToggle } from './ThemeToggle.svelte'
+export { default as ScrollArea } from './ScrollArea.svelte'
