+++ src/routes/+layout.svelte
@@ -9,8 +9,8 @@
   import { isMobileMenuOpen } from '$lib/stores/navigation'
   import { fade } from 'svelte/transition'
   import { ScrollArea } from '$lib/components'
-  import Button from '$lib/components/Button.svelte'
-  import ThemeToggle from '$lib/components/ThemeToggle.svelte'
+  import Button from '$lib/components/ui/Button.svelte'
+  import ThemeToggle from '$lib/components/ui/ThemeToggle.svelte'
   import Logo from '$lib/components/Logo.svelte'
   import md5 from 'md5'
   import { Menu, X, User, LogOut } from 'lucide-svelte'
