+++ src/routes/audit-logs/+page.svelte
@@ -7,8 +7,8 @@
    * what was changed, and when the changes occurred.
    */
   import type { PageData } from './$types'
-  import Feed from '$lib/components/Feed.svelte'
-  import type { FeedItem } from '$lib/components/Feed.svelte'
+  import Feed from '$lib/components/features/Feed.svelte'
+  import type { FeedItem } from '$lib/components/features/Feed.svelte'
   import md5 from 'md5'
 
   export let data: PageData
@@ -41,14 +41,14 @@
   }
 
   /**
-   * Formats changes from JSON string into a readable multi-line string
-   * @param {string} changesJson - JSON string containing change details
+   * Formats changes from JSON string or object into a readable multi-line string
+   * @param {string|Record<string, unknown>} changes - JSON string or object containing change details
    * @returns {string} Formatted string representing changes
    */
-  function formatChanges(changesJson: string) {
+  function formatChanges(changes: string | Record<string, unknown>): string {
     try {
-      const changes = JSON.parse(changesJson)
-      return Object.entries(changes)
+      const changesObj = typeof changes === 'string' ? JSON.parse(changes) : changes
+      return Object.entries(changesObj)
         .filter(([_, value]) => value !== undefined)
         .map(([key, value]) => {
           if (value && typeof value === 'object' && 'from' in value && 'to' in value) {
@@ -58,23 +58,25 @@
         })
         .join('\n')
     } catch (e) {
-      return changesJson
+      return typeof changes === 'string' ? changes : JSON.stringify(changes)
     }
   }
 
   /**
-   * Extracts deleted item information from JSON changes string
-   * @param {string} changesJson - JSON string containing change details
+   * Extracts deleted item information from JSON changes string or object
+   * @param {string|Record<string, unknown>} changes - JSON string or object containing change details
    * @returns {Record<string, any>|undefined} Object with deleted item properties or undefined
    */
-  function extractDeletedItem(changesJson: string): Record<string, any> | undefined {
+  function extractDeletedItem(
+    changes: string | Record<string, unknown>
+  ): Record<string, any> | undefined {
     try {
-      const changes = JSON.parse(changesJson)
-      if (changes.deleted && typeof changes.deleted === 'object') {
-        return changes.deleted
+      const changesObj = typeof changes === 'string' ? JSON.parse(changes) : changes
+      if (changesObj.deleted && typeof changesObj.deleted === 'object') {
+        return changesObj.deleted
       }
       const deletedValues: Record<string, any> = {}
-      for (const [key, value] of Object.entries(changes)) {
+      for (const [key, value] of Object.entries(changesObj)) {
         if (value && typeof value === 'object' && 'from' in value) {
           deletedValues[key] = value.from
         } else {
@@ -103,7 +105,9 @@
       },
       entityType: log.entityType,
       details: formatChanges(log.changes),
-      ...(log.actionType === 'DELETE' && { deletedItem: extractDeletedItem(log.changes) })
+      ...(log.actionType.toLowerCase() === 'delete' && {
+        deletedItem: extractDeletedItem(log.changes)
+      })
     })
   )
 </script>
