+++ src/routes/auth/login/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { enhance } from '$app/forms'
   import type { ActionData, PageData } from './$types'
-  import Input from '$lib/components/Input.svelte'
-  import Button from '$lib/components/Button.svelte'
+  import Input from '$lib/components/ui/Input.svelte'
+  import Button from '$lib/components/ui/Button.svelte'
   import Logo from '$lib/components/Logo.svelte'
 
   export let form: ActionData
