+++ src/routes/auth/register/+page.svelte
@@ -9,8 +9,8 @@
    */
   import { enhance } from '$app/forms'
   import type { ActionData } from './$types'
-  import Input from '$lib/components/Input.svelte'
-  import Button from '$lib/components/Button.svelte'
+  import Input from '$lib/components/ui/Input.svelte'
+  import Button from '$lib/components/ui/Button.svelte'
   import Logo from '$lib/components/Logo.svelte'
 
   /**
