+++ src/routes/proxy-hosts/+page.svelte
@@ -5,17 +5,17 @@
   import type { ProxyHostFormData } from './types'
   import type { SubmitFunction } from '@sveltejs/kit'
   import type { ComponentType, SvelteComponent } from 'svelte'
-  import Modal from '$lib/components/Modal.svelte'
-  import LoadingSpinner from '$lib/components/LoadingSpinner.svelte'
-  import ErrorAlert from '$lib/components/ErrorAlert.svelte'
-  import Input from '$lib/components/Input.svelte'
-  import Table from '$lib/components/Table.svelte'
+  import Modal from '$lib/components/layout/Modal.svelte'
+  import LoadingSpinner from '$lib/components/ui/LoadingSpinner.svelte'
+  import ErrorAlert from '$lib/components/ui/ErrorAlert.svelte'
+  import Input from '$lib/components/ui/Input.svelte'
+  import Table from '$lib/components/layout/Table.svelte'
   import { PencilLine, Trash2 } from 'lucide-svelte'
-  import Select from '$lib/components/Select.svelte'
-  import Toggle from '$lib/components/Toggle.svelte'
-  import Textarea from '$lib/components/Textarea.svelte'
-  import Button from '$lib/components/Button.svelte'
-  import ActionButton from '$lib/components/ActionButton.svelte'
+  import Select from '$lib/components/ui/Select.svelte'
+  import Toggle from '$lib/components/ui/Toggle.svelte'
+  import Textarea from '$lib/components/ui/Textarea.svelte'
+  import Button from '$lib/components/ui/Button.svelte'
+  import ActionButton from '$lib/components/features/ActionButton.svelte'
 
   export let data
   let form: ProxyHostFormData | null = null
